Added new example, modified old example, path fix
Things in this commit:
Created a new example that uses Particle.variable and Particle.publish
to expose/transmit data.
Fixed the import path in the original example so that it will compile
in Build.
Changed the original example’s count-based rate limiting to be directly
time-based, and easy to modify.
Fixed up some minor spelling things while I was in there.
Incremented version in spark.json for easy importing back into Build